fix: warnings -> https://github.com/purescript/documentation/blob/master/errors/HiddenConstructors.md
An export for Box hides data constructors but the type declares an instance of Data.Newtype.Newtype.
Such instance allows to match and construct values of this type, effectively making the constructors public.

An export for Alignment hides data constructors but the type declares an instance of Data.Generic.Rep.Generic.
Such instance allows to match and construct values of this type, effectively making the constructors public.